+++ R/get_data.R
@@ -1975,6 +1975,15 @@ get_data.svyglm <- function(x, source = "mf", ...) {
   }
 }
 
+#' @export
+get_data.svyolr <- get_data.svyglm
+
+#' @export
+get_data.svycoxph <- get_data.svyglm
+
+#' @export
+get_data.svysurvreg <- get_data.svyglm
+
 
 # mfx models ------------------------------------------------------