+++ R/get_weights.R
@@ -142,6 +142,15 @@ get_weights.survey.design2 <- get_weights.survey.design
 #' @export
 get_weights.svyglm <- get_weights.survey.design
 
+#' @export
+get_weights.svyolr <- get_weights.survey.design
+
+#' @export
+get_weights.svycoxph <- get_weights.survey.design
+
+#' @export
+get_weights.svysurvreg <- get_weights.survey.design
+
 
 #' @export
 get_weights.list <- function(x, remove_na = FALSE, null_as_ones = FALSE, ...) {