+++ R/is_model_supported.R
@@ -116,7 +116,7 @@ supported_models <- function() {
     "Sarlm", "scam", "selection", "sem", "semLm", "semLme", "SemiParBIV", "serp",
     "slm", "speedlm", "speedglm", "stanfit", "stanmvreg", "stanreg", "summary.lm",
     "survfit", "survreg", "svy_vglm", "svychisq", "svyglm", "svyolr", "svy2lme",
-    "seqanova.svyglm", "sdmTMB", "systemfit",
+    "svycoxph", "seqanova.svyglm", "sdmTMB", "systemfit", "svysurvreg",
 
     # t ----------------------------
     "t1way", "tobit", "trimcibt", "truncreg",